facttrace/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
crewai>=0.80.0
